requirements.txt
streamlit>=1.31.0
pypdf>=4.0.0
python-docx>=1.1.0
chromadb>=0.4.22
python-dotenv>=1.0.0
gigachat>=0.2.0
sentence-transformers>=2.2.0
requests>=2.31.0